+++ modules/core/src/main/scala/fs2/kafka/package.scala
@@ -16,13 +16,9 @@
 
 package fs2
 
-import java.util.concurrent.{Executors, ThreadFactory}
-
 import cats.Applicative
 import cats.effect._
 import cats.syntax.traverse._
-
-import scala.concurrent.ExecutionContext
 import scala.concurrent.duration.FiniteDuration
 
 package object kafka {
@@ -428,226 +424,6 @@ package object kafka {
     */
   def consumerStream[F[_]](implicit F: ConcurrentEffect[F]): ConsumerStream[F] =
     new ConsumerStream[F](F)
-
-  /**
-    * Creates a new `ExecutionContext` backed by a single thread.
-    * This is suitable for use with a single `KafkaConsumer`, and
-    * is default unless a different context is set using
-    * [[ConsumerSettings#withExecutionContext]].<br>
-    * <br>
-    * If you already have an `ExecutionContext` for blocking code,
-    * then you might prefer to use that over explicitly creating
-    * one with this function.<br>
-    * <br>
-    * The thread created by this function will be of type daemon,
-    * and the `Resource` context will automatically shutdown the
-    * underlying `Executor` as part of finalization.<br>
-    * <br>
-    * You might prefer `consumerExecutionContextStream`, which is
-    * returning a `Stream` instead of `Resource`. For convenience
-    * when working together with `Stream`s.
-    */
-  def consumerExecutionContextResource[F[_]](
-    implicit F: Sync[F]
-  ): Resource[F, ExecutionContext] =
-    consumerExecutionContextResource(1)
-
-  /**
-    * Creates a new `ExecutionContext` backed by the specified number
-    * of `threads`. This is suitable for use with the same number of
-    * `KafkaConsumer`s, and can be specified when creating a
-    * [[ConsumerSettings]] instance.<br>
-    * <br>
-    * If you already have an `ExecutionContext` for blocking code,
-    * then you might prefer to use that over explicitly creating
-    * one with this function.<br>
-    * <br>
-    * The threads created by this function will be of type daemon,
-    * and the `Resource` context will automatically shutdown the
-    * underlying `Executor` as part of finalization.<br>
-    * <br>
-    * You might prefer `consumerExecutionContextStream`, which is
-    * returning a `Stream` instead of `Resource`. For convenience
-    * when working together with `Stream`s.
-    */
-  def consumerExecutionContextResource[F[_]](threads: Int)(
-    implicit F: Sync[F]
-  ): Resource[F, ExecutionContext] =
-    executionContextResource("fs2-kafka-consumer", threads)
-
-  /**
-    * Creates a new `ExecutionContext` backed by a single thread.
-    * This is suitable for use with a single `KafkaProducer`,
-    * and is default unless a different context is set using
-    * [[ProducerSettings#withExecutionContext]].<br>
-    * <br>
-    * If you already have an `ExecutionContext` for blocking code,
-    * then you might prefer to use that over explicitly creating
-    * one with this function.<br>
-    * <br>
-    * The thread created by this function will be of type daemon,
-    * and the `Resource` context will automatically shutdown the
-    * underlying `Executor` as part of finalization.<br>
-    * <br>
-    * You might prefer `producerExecutionContextStream`, which is
-    * returning a `Stream` instead of `Resource`. For convenience
-    * when working together with `Stream`s.
-    */
-  def producerExecutionContextResource[F[_]](
-    implicit F: Sync[F]
-  ): Resource[F, ExecutionContext] =
-    producerExecutionContextResource(1)
-
-  /**
-    * Creates a new `ExecutionContext` backed by the specified number
-    * of `threads`. This is suitable for use with the same number of
-    * `KafkaProducer`s, and can be specified when creating a
-    * [[ProducerSettings]] instance.<br>
-    * <br>
-    * If you already have an `ExecutionContext` for blocking code,
-    * then you might prefer to use that over explicitly creating
-    * one with this function.<br>
-    * <br>
-    * The threads created by this function will be of type daemon,
-    * and the `Resource` context will automatically shutdown the
-    * underlying `Executor` as part of finalization.<br>
-    * <br>
-    * You might prefer `producerExecutionContextStream`, which is
-    * returning a `Stream` instead of `Resource`. For convenience
-    * when working together with `Stream`s.
-    */
-  def producerExecutionContextResource[F[_]](threads: Int)(
-    implicit F: Sync[F]
-  ): Resource[F, ExecutionContext] =
-    executionContextResource("fs2-kafka-producer", threads)
-
-  /**
-    * Creates a new `ExecutionContext` backed by a single thread.
-    * This is suitable for use with a single `KafkaAdminClient`,
-    * and is default unless a different context is set using
-    * [[AdminClientSettings#withExecutionContext]].<br>
-    * <br>
-    * If you already have an `ExecutionContext` for blocking code,
-    * then you might prefer to use that over explicitly creating
-    * one with this function.<br>
-    * <br>
-    * The thread created by this function will be of type daemon,
-    * and the `Resource` context will automatically shutdown the
-    * underlying `Executor` as part of finalization.<br>
-    * <br>
-    * You might prefer `adminClientExecutionContextStream`, which is
-    * returning a `Stream` instead of `Resource`. For convenience
-    * when working together with `Stream`s.
-    */
-  def adminClientExecutionContextResource[F[_]](
-    implicit F: Sync[F]
-  ): Resource[F, ExecutionContext] =
-    adminClientExecutionContextResource(1)
-
-  /**
-    * Creates a new `ExecutionContext` backed by the specified number
-    * of `threads`. This is suitable for use with the same number of
-    * `KafkaAdminClient`s, and can be specified when creating an
-    * [[AdminClientSettings]] instance.<br>
-    * <br>
-    * If you already have an `ExecutionContext` for blocking code,
-    * then you might prefer to use that over explicitly creating
-    * one with this function.<br>
-    * <br>
-    * The threads created by this function will be of type daemon,
-    * and the `Resource` context will automatically shutdown the
-    * underlying `Executor` as part of finalization.<br>
-    * <br>
-    * You might prefer `adminClientExecutionContextStream`, which is
-    * returning a `Stream` instead of `Resource`. For convenience
-    * when working together with `Stream`s.
-    */
-  def adminClientExecutionContextResource[F[_]](threads: Int)(
-    implicit F: Sync[F]
-  ): Resource[F, ExecutionContext] =
-    executionContextResource("fs2-kafka-admin-client", threads)
-
-  private[this] def executionContextResource[F[_]](
-    name: String,
-    threads: Int
-  )(implicit F: Sync[F]): Resource[F, ExecutionContext] =
-    Resource
-      .make {
-        F.delay {
-          Executors.newFixedThreadPool(
-            threads,
-            new ThreadFactory {
-              override def newThread(runnable: Runnable): Thread = {
-                val thread = new Thread(runnable)
-                thread.setName(s"$name-${thread.getId}")
-                thread.setDaemon(true)
-                thread
-              }
-            }
-          )
-        }
-      }(executor => F.delay(executor.shutdown()))
-      .map(ExecutionContext.fromExecutor)
-
-  /**
-    * Like `consumerExecutionContextResource`, but returns a `Stream`
-    * rather than a `Resource`. This is for convenience when working
-    * together with `Stream`s.
-    */
-  def consumerExecutionContextStream[F[_]](
-    implicit F: Sync[F]
-  ): Stream[F, ExecutionContext] =
-    Stream.resource(consumerExecutionContextResource)
-
-  /**
-    * Like `consumerExecutionContextResource`, but returns a `Stream`
-    * rather than a `Resource`. This is for convenience when working
-    * together with `Stream`s.
-    */
-  def consumerExecutionContextStream[F[_]](threads: Int)(
-    implicit F: Sync[F]
-  ): Stream[F, ExecutionContext] =
-    Stream.resource(consumerExecutionContextResource(threads))
-
-  /**
-    * Like `producerExecutionContextResource`, but returns a `Stream`
-    * rather than a `Resource`. This is for convenience when working
-    * together with `Stream`s.
-    */
-  def producerExecutionContextStream[F[_]](
-    implicit F: Sync[F]
-  ): Stream[F, ExecutionContext] =
-    Stream.resource(producerExecutionContextResource)
-
-  /**
-    * Like `producerExecutionContextResource`, but returns a `Stream`
-    * rather than a `Resource`. This is for convenience when working
-    * together with `Stream`s.
-    */
-  def producerExecutionContextStream[F[_]](threads: Int)(
-    implicit F: Sync[F]
-  ): Stream[F, ExecutionContext] =
-    Stream.resource(producerExecutionContextResource(threads))
-
-  /**
-    * Like `adminClientExecutionContextResource`, but returns a `Stream`
-    * rather than a `Resource`. This is for convenience when working
-    * together with `Stream`s.
-    */
-  def adminClientExecutionContextStream[F[_]](
-    implicit F: Sync[F]
-  ): Stream[F, ExecutionContext] =
-    Stream.resource(adminClientExecutionContextResource)
-
-  /**
-    * Like `adminClientExecutionContextResource`, but returns a `Stream`
-    * rather than a `Resource`. This is for convenience when working
-    * together with `Stream`s.
-    */
-  def adminClientExecutionContextStream[F[_]](threads: Int)(
-    implicit F: Sync[F]
-  ): Stream[F, ExecutionContext] =
-    Stream.resource(adminClientExecutionContextResource(threads))
 
   /**
     * Creates a new [[KafkaProducer]] in the `Resource` context,
